Factory Design Pattern few changes and Test

+++ src/Exercise/FactoryDesignPattern/GetPlan.java
@@ -8,7 +8,7 @@ class GetPlan {
          System.out.println(" For Domestic plan select 1 and for Commercial plan select 2 ");
          int ab =input.nextInt();
          System.out.println("Enter electricity units : ");
-         int units = input.nextInt();
+         double units = input.nextDouble();
          if(ab==1){
              System.out.println("Domestic Plan");
             DomesticPlan dp= new DomesticPlan();
